Return int from SqlTimestamp.getPicosOfMicros()

+++ presto-spi/src/main/java/io/prestosql/spi/type/SqlTimestamp.java
@@ -87,7 +87,7 @@ public final class SqlTimestamp
         return epochMicros;
     }
 
-    public long getPicosOfMicros()
+    public int getPicosOfMicros()
     {
         return picosOfMicros;
     }
